Beat & Note Array experiments
Phil created a "drum machine" in Sonic Pi with arrays (based on 16 beat drum machines) and a regular "sleep" between each beat. ":z" = a reference to "no sample" (aka silence). 
The note array test plays with an array of notes & an array of times using "tick" as a variable (if you reference "tick" more than once (even using "puts" to see what the value of tick is, it increments the tick, so 1, 2, 3, 4 etc can turn into 1, 5, 9, 13 etc.)